More updates

+++ engine/Location.js
@@ -1,91 +1,106 @@
 const Base = require("./Base");
 const Player = require("./Player");
 const Collection = require("./Collection");
-const Utility = require("./Utility");
+const utility = require("./utility");
 const Mob = require("./Mob");
-const Monster = require("./Monster");
 const Action = require("./Action");
-const Battle = require("./Battle");
+const { Message, VoiceChannel } = require("discord.js");
 
 /**
  * Represents a location
  * @extends {Base}
+ * @public
  * @param {World} world
  * @param {Object} options
  */
 class Location extends Base {
+	static LOCATION_GENERATION_ERROR = class LOCATION_GENERATION_ERROR extends Error { };
+	static LOCATION_UNGENERATION_ERROR = class LOCATION_UNGENERATION_ERROR extends Error { };
+	static LOCATION_ATTACHMENT_ERROR = class LOCATION_ATTACHMENT_ERROR extends Error { };
+	static LOCATION_NARRATION_ERROR = class LOCATION_NARRATION_ERROR extends Error { };
 	constructor(world, options = {}) {
-		if (!options.locations)
-			options.locations = new Collection(Location);
+		if (!options.name) options.name = "A Location";
 
 		super(world);
 
 		this.battle = null;
 		/**
 		 * Indicates whether this location has been generated
+		 * @public
 		 * @type {Boolean}
 		 */
 		this.generated = false;
 		/**
 		 * All mobs currently at this location
+		 * @public
 		 * @type {Collection<Mob>}
 		 */
 		this.mobs = new Collection(Mob);
 		/**
 		 * All actions taken at this location
+		 * @public
 		 * @type {Collection<Action>}
 		 */
 		this.actions = new Collection(Action);
 		/**
 		 * The role associated with this location
+		 * @public
 		 * @type {Role}
 		 */
 		this.role = null;
 		/**
 		 * The category associated with this location
+		 * @public
 		 * @type {CategoryChannel}
 		 */
 		this.category = null;
 		/**
 		 * The text channel associated with this location
+		 * @public
 		 * @type {TextChannel}
 		 */
 		this.textChannel = null;
 		/**
 		 * The voice channel associated with this location
+		 * @public
 		 * @type {VoiceChannel}
 		 */
 		this.voiceChannel = null;
 		/**
 		 * The channel used as a separator between the button channels and the voice/text channels
+		 * @public
 		 * @type {TextChannel}
 		 */
 		this.spacerChannel = null;
 		/**
 		 * The name of the location
+		 * @public
 		 * @type {String}
 		 */
 		this.name = options.name;
 		/**
 		 * Any and all locations attached to this location
+		 * @public
 		 * @type {Collection<Location>}
 		 */
-		this.locations = options.locations;
+		this.locations = new Collection(Location);
 		/**
 		 * Contains all the voice channel buttons for moving between locations
-		 * @type {Collection}
+		 * @public
+		 * @type {Collection<VoiceChannel>}
 		 */
-		this.buttons = new Collection();
+		this.buttons = new Collection(VoiceChannel);
 		this.world.locations.add(this);
 	}
 	/**
 	 * Creates the role and channels for this location and links the associated locations to the newly created button channels
 	 * @async
-	 * @returns {void}
+	 * @public
+	 * @returns {Promise<void>}
 	 */
 	async generate() {
 		if (this.generated)
-			throw new Error("Location must not be generated to be generated");
+			throw new LOCATION_GENERATION_ERROR("Location must not be generated to be generated");
 		this.role = await this.guild.roles.create({
 			data: {
 				name: this.name
@@ -128,20 +143,7 @@ class Location extends Base {
 			});
 		}
 		for (let location of this.locations) {
-			await this._addLocationButton(location);
-		}
-		for (let mob of this.mobs) {
-			if (mob[1] instanceof Player) {
-				await mob[1].guildMember.roles.add(this.role);
-				if (mob[1].guildMember.voice.speaking != null) {
-					await mob[1].guildMember.voice.setChannel(this.voiceChannel);
-				}
-				await this.textChannel.send({
-					embed: {
-						description: `${mob[1].name} enters.`
-					}
-				});
-			}
+			await this._addLocationButton(location[1]);
 		}
 		this.generated = true;
 		await this.emit("generated");
@@ -149,23 +151,29 @@ class Location extends Base {
 	/**
 	 * Reverses the effects of the `generate()` method
 	 * @async
-	 * @returns {Promise}
+	 * @public
+	 * @returns {Promise<void>}
 	 */
 	async ungenerate() {
-		if (!this.generated)
-			throw new Error("Location must be generated to ungenerate");
+		if (!this.generated) throw new LOCATION_UNGENERATION_ERROR("Location must be generated to ungenerate");
 		for (let button of this.buttons) {
 			await button[1].delete();
 		}
 		await this.textChannel.delete();
 		await this.voiceChannel.delete();
-		if (this.spacerChannel)
-			await this.spacerChannel.delete();
+		if (this.spacerChannel) await this.spacerChannel.delete();
 		await this.category.delete();
 		await this.role.delete();
 		this.generated = false;
 		await this.emit("ungenerated");
 	}
+	/**
+	 * Registers an action at this location
+	 * @async
+	 * @private
+	 * @param {Action} action 
+	 * @returns {Promise<void>}
+	 */
 	async _registerAction(action) {
 		action.location = this;
 		this.actions.add(action);
@@ -180,6 +188,13 @@ class Location extends Base {
 		});
 		await this.emit("actionTaken", action);
 	}
+	/**
+	 * Adds a VoiceChannel button for a location to this location
+	 * @async
+	 * @private
+	 * @param {Location} location
+	 * @returns {Promise<void>}
+	 */
 	async _addLocationButton(location) {
 		let button = await this.guild.channels.create(location.name, {
 			type: "voice",
@@ -192,7 +207,7 @@ class Location extends Base {
 				id: this.guild.roles.everyone,
 				deny: ["SEND_MESSAGES", "CONNECT", "SPEAK", "VIEW_CHANNEL"]
 			}],
-			position: count
+			position: this.category.children.size + 1
 		});
 		this.buttons.add(button);
 		let world = location.world;
@@ -213,36 +228,75 @@ class Location extends Base {
 		return button;
 	}
 	/**
+	 * Adds a mob to this location
+	 * @async
+	 * @private
+	 * @param {MobResolvable} mobResolvable 
+	 * @returns {Promise<void>}
+	 */
+	async _addMob(mobResolvable) {
+		let mob = this.world.mobs.resolve(mobResolvable);
+		this.mobs.add(mob);
+		if (this.generated) {
+			await this.textChannel.send({
+				embed: {
+					description: `${mob.name} enters.`
+				}
+			});
+		}
+		await this.emit("mobJoined", mob);
+	}
+	/**
+	 * Removes a mob from this location
+	 * @async
+	 * @private
+	 * @param {MobResolvable} mobResolvable 
+	 * @returns {Promise<void>}
+	 */
+	async _removeMob(mobResolvable) {
+		let mob = this.world.mobs.resolve(mobResolvable);
+		this.mobs.remove(mob);
+		if (this.generated) {
+			await this.textChannel.send({
+				embed: {
+					description: `${mob.name} leaves.`
+				}
+			});
+		}
+		await this.emit("mobLeft", mob);
+	}
+	/**
 	 * Places a location next to this one
 	 * @param {LocationResolvable} locationResolvable - The location to attach
 	 * @async
-	 * @returns {Promise}
+	 * @public
+	 * @returns {Promise<void>}
 	 */
 	async attach(locationResolvable) {
-		if (!locationResolvable)
-			throw new Error(`Requires a location.`);
-		let location = this.world.locations.resolve(location);
+		if (!locationResolvable) throw new LOCATION_ATTACHMENT_ERROR('Requires a location.');
+		let location = this.world.locations.resolve(locationResolvable);
 		this.locations.add(location);
 		if (this.generated) {
-			await this._addLocationButton(this[buttonString], this[direction]);
+			await this._addLocationButton(location);
 		}
+		await this.emit("locationAttached", location);
 	}
 	/**
 	 * Sends a narration message to the `textChannel` property channel
 	 * @param {String} message - The message text
 	 * @async
+	 * @public
 	 * @returns {Promise<Message>}
 	 */
 	async narrate(message) {
-		if (!this.generated)
-			throw new Error(`Location not generated.`);
-		if (!message)
-			throw new Error(`Requires a string.`);
+		if (!this.generated) throw new LOCATION_NARRATION_ERROR('Location not generated.');
+		if (!message) throw new Error('Requires a string.');
 		let actualMessage = await this.textChannel.send({
 			embed: {
 				description: message
 			}
 		});
+		await this.emit("narration", actualMessage);
 		return actualMessage;
 	}
 }
@@ -257,6 +311,18 @@ module.exports = Location;
 /**
  * Emitted when this location is ungenerated
  * @event Location#ungenerated
+ */
+
+/**
+ * Emitted when another location is attached to this one
+ * @event Location#locationAttached
+ * @param {Location} [location]
+ */
+
+/**
+ * Emitted when a narration message is sent
+ * @event Location#narration
+ * @param {Message} [message]
  */
 
 /**
